Create PoolScopeTreeProp and wire it with Primitive API
Summary:
This is a part of [Litho Custom Scoped Content Pools](https://docs.google.com/document/d/14vBVBTu8RIfaF_V4HV4rZAHlTmF1uwMdnd94SlgpQWM/edit?usp=sharing) work.

In this diff we're introducing an internal `PoolScopeTreeProp` that will be used to propagate `PoolScope` to the children via a `TreeProp`.
We're also exposing an API for providing `PoolScope` to `ComponentTree` and adding logic for reading and passing it automatically to the `Primitive`.

Reviewed By: adityasharat

Differential Revision: D64236155

fbshipit-source-id: cfcbf2802b496ca97a364b6c5b74738d52e08a2e