Fix transition tests
Summary:
they were broken for a while.
After fixing fresco issue here:
D64691550
the next error was: snapActivity for v2 screenshots can only be invoked in the main thread

I tried adding looper, but then i had issues with dex:
Cannot fit requested classes in a single dex file
and splitting didnt solve anything...

so i found "UiThreadTest" annotation but then our ComponentActivityTestRule required us not to be on main thread.

I modified this class and provided variant that allowed setting the component on main thread and was able to move a bit with execution, but then clicking on views was crashing with the same error that we cannot be on main thread.

after fighting for a while with this I decided to create our custom litho rule that is using Screenshotter directly without the wrapper for v2.

If in the future testing team will want to delete Screenshotter, i hope they will help us to migrate those tests or rewrite them properly.

Reviewed By: zielinskimz

Differential Revision: D64760718

fbshipit-source-id: 72b83f3e90690925ac702e6783bd59a7b67347eb