Fix content description bug
Summary:
This is the exact thing we shipped back in the day and got reverted due to sev.
But now after investigating we were able to find that the talkback is not broken due to the fix but its broken due to infinite loop of updates that NewsFeedFeedUnitComponentSpec is doing.

Reviewed By: pentiumao

Differential Revision: D64952646

fbshipit-source-id: 11ae53ce7aee12e342acec23f862f2958cc3bf62