Add support for MountableSpan interface
Summary:
This diff introduces the `MountableSpan` interface that can be added to custom spans to receive callback with the hosting View when the component is mounted or unmounted.

This approach was also discussed in the scope of the [Composable and Extensible span RFC for Bloks](https://docs.google.com/document/d/1EKlAx9dMWsrltmmN1Ge9NfsXSycSXKwCLcmoCoxbwlE/edit?tab=t.0#bookmark=id.l3rxz5cm5xj2) but it was never implemented.

Now this interface is used for the GenAI use case implemented in the next diff.

Reviewed By: zielinskimz

Differential Revision: D64940916

fbshipit-source-id: 12b4f1647e6e7f88f54078f598d9522a0c39589c