Add ScreenReaderFocusable common prop
Summary:
This adds support for `setScreenReaderFocusable` to Litho

https://developer.android.com/reference/android/view/View#setScreenReaderFocusable(boolean)

This is needed for accessibility usecases where we want to differentiate between focus for talkback and focus for keyboard navigation.

The most common case of this we have run into is when applying `accessibilityHeading` to a text view in a view group. The default behavior will group all the view group children together and read them out at once, ignoring the heading label.

The wiki to this is to apply `focusable` to the element
https://www.internalfb.com/intern/wiki/Accessibility_Patterns/Focus/Android_Automatic_Focus_Grouping/#the-solution

However, as noted elsewhere in the wiki, this will _also_ make the element focusable for keyboard navigation. This is a bug, since keyboard should only navigate between actionable elements.

The solution shared elsewhere in the wiki is to use `screenReaderFocusable(true)` to control this. However this property was not exported through litho, hence this diff

https://www.internalfb.com/intern/wiki/Accessibility_Patterns/Keyboard_Focus/Unnecessary_Focus/#android

See discussion of this issue in T203078327 and https://fb.workplace.com/groups/3615245781855602/posts/8517755938271204/?comment_id=8522609674452497&reply_comment_id=8524605904252874

Reviewed By: adityasharat

Differential Revision: D64933701

fbshipit-source-id: 3a61f503a3015ef24977859b7433ba605a8fe3e3