Make activity section header in Community Notifications Settings not focusable
Summary:
As part of the Android Accessibility Wave, we want to remove **keyboard focus** for the "Activity" header of the Community Notification Settings

This is because keyboard focus should only focus on relevant screen content. Since the "Activity" heading is considered a "non-interactive" element, we want to remove it from keyboard focus.

This diff:
- Replaces `focusable` with `screenReaderFocusable` for **MigSectionHeader**
- Adds `focusable` to **MigSectionHeaderCTASpec**. That way, the header only gets focusability when it's interactive
- Add new test cases to test that the header is screen reader focusable and is only focusable if it's interactive

Differential Revision: D64793278

fbshipit-source-id: cf6fb3243f8e48ecd3905390ac23aa03e690f6ab