Make mount order is also from leaf node
Summary:
## Context

We should update `MountItem` from leaf node as well, otherwise, we could get incorrect state for each node and end up doing unnecessary mount/unmount.

e.g. we should mount a Host if it has mounted children, but we'll end up unmounting it if we process in the normal order.

Reviewed By: fbcbl

Differential Revision: D65611600

fbshipit-source-id: bec4264baf3f20ac44f199b2eaedfcb379d9750f