Add ContextThemeWrapper instead of inflating the layout
Summary:
Primitive vertical scroll was crashing for bottom sheet.
Fabio findings:
the vertical spec inflates an XML which has the android:scrollbars=vertical.
This causes a scrollbar related initialization code to run inside View.java (https://fburl.com/f02mlsef) . As there is no good way to call this from public methods, then we need to create the LithoScrollView with a style that also adds the "scrollbar=vertical"

I tried adding the style as a construction param inside the LithoScrollView, but we still were crashing.
So adding this to unblock the QE.

Reviewed By: zielinskimz

Differential Revision: D66098277

fbshipit-source-id: edd7cb0a9c8b40e159edbb40bd369312349a7bec