drop layout context from render tree
Summary:
# What?
Storing the layout context in the render tree means that we hold references to the read and write cache per generation which means that we essentially have a chain of caches as new renders happen. Caches can themselves reference layout contexts if they have render trees in them – for example collection cells. This causes caches from older generations to never get GC'd until the screen or integration is destroyed and sometimes never GC'd if the surface leaks `BloksContext` beyond destruction. This eventually results in an OOM.

# This Diff
We required the layout context in the render tree in the first place to gain access to the render context for signal evaluations. This diff just threads the render context through instead of holding on to the layout caches by way of the layout context. This means that for a given surface, memory should never accumulate steeply. Even in the worst case, where `BloksContext` is leaked, memory utilization from the PoV of `LayoutContext`s and `LayoutCache`s held in memory should stay the same.

# What's the cost?
With this diff, any `RenderTree`s in a given generation of render should hold a reference (~4-8 bytes) to `BloksContext` (of which there is only ever one instance for a screen) and significantly better than holding on to all the layout contexts that have ever existed on screen

Reviewed By: AdamMcCormick

Differential Revision: D66518227

fbshipit-source-id: 22c5b6cfd02da826ed3a5002a70db45d8e2fe5d3