Handle infinite measuring
Summary:
the new crashes for primitive happens due to the fact that we are getting:
```
sizeConstraints.maxHeight = infinity, min = 0
```
when fillViewport is true we are doing:
```
if (fillViewport) {
          sizeConstraints.copy(
              minHeight = sizeConstraints.maxHeight,
              maxHeight = sizeConstraints.maxHeight,
          )
```
which gives: infinity, infinity so we crash.

in spec we are creating a wrapper and we are setting:
SizeSpec.getSize(heightSpec)
which in this case is 0

changing to the new logic we would get:

```
SizeConstraints[minWidth = 1080, maxWidth = 1080, minHeight = 0, maxHeight = Infinity]
```
which is what we want.

Reviewed By: adityasharat

Differential Revision: D66538510

fbshipit-source-id: 13915447ec95c16b40a3d3a57cf15f53c9480bb7