Measurement fix
Summary:
I published the diff that started collecting errors if we would ship primitive vertical scroll: D66098306
and we got some measurements errors like:

```
com.facebook.litho.LithoMetadataExceptionWrapper: Real Cause at
<cls>com.facebook.litho.widget.ExperimentalVerticalScroll</cls> => java.lang.IllegalArgumentException:
minHeight must be <= 8190, but was: 1073741760. Components this big may affect performance and lead to
out of memory errors. minWidth=1048, maxWidth=1048, minHeight=1073741760, maxHeight=2147483647
```
With zielinskimz help this means:
minWidth=1048, maxWidth=1048 -> so the width is EXACT

minHeight=1073741760, maxHeight=2147483647 -> so we have a RANGE as min != max

(decoding logic for SizeConstraints here https://www.internalfb.com/code/fbsource/fbandroid/libraries/rendercore/rendercore/src/main/java/com/facebook/rendercore/SizeConstraints.kt?lines=385-385 )

and then in our Primitive we are crashing when we are trying to create size constraints:
https://www.internalfb.com/code/fbsource/[1f6adbdd0beb69f931b12bad10ceff29c0f96610]/fbandroid/libraries/components/litho-widget/src/main/java/com/facebook/litho/widget/ExperimentalVerticalScroll.kt?lines=185-188

so we are getting size constraints were maxHeight is 1073741760
and we set here minHeight = maxHeight (1073741760)
and maxHeight for infinity

trying something like this is not working:
```
sizeConstraints.copy(
              minHeight = min(sizeConstraints.maxHeight, sizeConstraints.MaxPossibleHeightValue),
              maxHeight = SizeConstraints.Infinity,
          )
```

as sizeConstraints.MaxPossibleHeightValue is still 1073741760 so it didn't change anything...

and it seems the error happens when the available space is infiniti so we flip it to max.

Reviewed By: kingsleyadio

Differential Revision: D66535003

fbshipit-source-id: 1515ff9747179806f75c379296a204184488d0c8