Change the max from infinity to max
Summary:
In the previous version of the scroll we were getting a screenshot test failure:
https://www.internalfb.com/intern/testinfra/diagnostics/15481123779000316.844425031843813.1733314395/
After adding some logs it looks like our scroll is getting:  height 175px EXACT
in Vertical Scroll Component, child is always measured wth heightSpec [0, UNSPECIFIED]
https://www.internalfb.com/code/fbsource/[2bdf4aac903853ca06cbd3fc7513f97af4bd8bd3]/fbandroid/libraries/components/litho-widget/src/main/java/com/facebook/litho/widget/VerticalScrollComponentSpec.java?lines=183
so it can have anything as its height.
BUT, in Spec code we are wrapping our child and set minimum Height (in our test case 175)
https://www.internalfb.com/code/fbsource/[2bdf4aac903853ca06cbd3fc7513f97af4bd8bd3]/fbandroid/libraries/components/litho-widget/src/main/java/com/facebook/litho/widget/VerticalScrollComponentSpec.java?lines=175-180
so we measure the child with minHeight = 175 and max as unspecified. (we respect minimum here and max can be anything)

In Primitive our sizeConstraints are minHeight=175, maxHeight=175 (EXACT) and we change them to
minHeight=175, maxHeight=Infinity (which makes sense, as we just try to say to the child that it can take anything it needs as its height)
https://www.internalfb.com/code/fbsource/[7b85e649703658f909efe30c254bc6f46889ad82]/fbandroid/libraries/components/litho-widget/src/main/java/com/facebook/litho/widget/ExperimentalVerticalScroll.kt?lines=208-211

The issue here is that we are doing NestedLithoTree.layout and we are giving our size constraints where we ignore the minimum height. We pass them down and convert to Measure Spec here:
https://www.internalfb.com/code/fbsource/[2daeffc7727f6d37bdf33fc4c1153e99df9cb86c]/fbandroid/libraries/components/litho-core/src/main/java/com/facebook/litho/LithoYogaLayoutFunction.kt?lines=77-78

And with this conversion, our minHeight is getting lost, our SizeConstraints[minHeight=175, maxHeight=Infinity] is turned into  MeasureSpec [0, UNSPECIFIED] (Unspecified as the maxHeight = infinity)

We don't have a way to tell Yoga that we have a minimum height that we need to respect.

So this diff will make our tests pass, but the measuring of the scroll is still not right. We will tell our child to be size 175, just like its parent, but because it is a scroll view it can be of any size and we should be able to just scroll it.

Some of our primitive components are doing hack like:
https://www.internalfb.com/code/fbsource/[64a725e54c705dc413c0ceb0d0efda0e25575db0]/fbandroid/java/com/instagram/litho/ui/spinnerimage/SpinnerImageComponent.kt?lines=54-61
But we dont know if this would even work for our vertical scroll

with huge help of zielinskimz it was possible to come to this conclusion

Reviewed By: zielinskimz

Differential Revision: D66758287

fbshipit-source-id: d928a170c864b5d046f7e32e44219e8f0abffb72