Introduce `crossAxisWrapContent` for RecyclerBinder
Summary:
## Context

I noticed that `BetterRecyclerCollectionComponent::canMeasureRecycler` should be a config and never changes after RecyclerBinder's initialization.

I traced back and noticed that `canMeasure` was introduced by D5020425, the purpose of the change was re-measuring the list.

And later on, we added support for **WRAP_CONTENT** of RecyclerBinder(D7687935), which should be exactly the same effect of `canMeasure`, but we didn't remove `canMeasure`.

In addition to that, we introduced `canMeasureRecycler` and `setCanMeasure` in D12890980, then cleaned it up with D12890980, which was addressing the issue with H-Scroll.

https://www.internalfb.com/code/fbsource/[04ce569f14ddf34b65d9870a744f85b4327bf082]/fbandroid/libraries/components/litho-widget/src/main/java/com/facebook/litho/widget/Binder.java?lines=99-103

After that, we introduced `CrossAxisWrapMode.canMeasureRecycler` for `LazyCollection`(D31688566), and reused this API(`setCanMeasure`):

https://www.internalfb.com/code/fbsource/[3202f89cb4f3555e16a3e3fda56eff3fb770577f]/fbandroid/libraries/components/litho-widget-kotlin/src/main/kotlin/com/facebook/litho/widget/collection/CollectionRecyclerSpec.kt?lines=246

I think this API is a bit confusing and can lead to bad usages, which would be better to rename it as `crossAxisWrapContent`. In this diff I primarily focus on replacing `canMeasureRecycler` with `crossAxisWrapContent`, and adding it into RecyclerBinderConfig.

Given removing `canMeasureRecycler` can break a lot of callsites, so wanted to request an early feedback here, please igore any build error.

## Proposal

1. To support `WRAP_CONTENT` for main axis, we should use:

```
RecyclerBinderConfig.mainAxisWrapContent(true) // now it is ::wrapContent
```

2. To support `WRAP_CONTENT` for cross axis, we should use:

```
RecyclerBinderConfig.crossAxisWrapContent(true)
```

Reviewed By: astreet

Differential Revision: D66450825

fbshipit-source-id: 7893d1c0514ae4172983a7bcfe7c9105b06290ef