Add debugger class that tracks animatedState lifecycles
Summary:
# Context
We are trying to resolve the crash associated to this task (T209041862) which happens due to an inconsistency between `initialStatesToRestore` and `animatedStates` in Litho's `TransitionManager`. Concretely, there is a missing `animatedState` for a given entry on `initialStatesToRestore`.

# This diff
In this diff we introduce this `AnimationInconsistencyDebugger`, which tracks whenever an `AnimatedState` is added or removed from `animatedStates`. Then on the custom exception we collect this data, so that we can try to understand when the missing `animatedState` was removed (if ever was removed).

Below there is a high level overview of the points I identify that we add or remove (in red) animation states. Therefore, I believe that one of these flows is triggered somehow before the initial states are restored, and that is where the bug is (and hopefully I can identify a pattern after this is landed).

 {F1973763534}

Reviewed By: zielinskimz

Differential Revision: D67034475

fbshipit-source-id: 79ac843153d12cfa620c0329da3c3bc638533566