Add ability to clean up initial states after removing animation state for undeclared transitions
Summary:
# Context
We are trying to resolve the crash associated to this task (T209041862) which happens due to an inconsistency between `initialStatesToRestore` and `animatedStates` in Litho's `TransitionManager`. Concretely, there is a missing `animatedState` for a given entry on `initialStatesToRestore`.

# This diff
We have identified two sections of the code where we remove "unused" `AnimationStates` right after the `initialStatesToRestore` are populated, and before the  actual transitions are ran.

This means that there is space for that cleanup to remove an animation that somehow is associated to an `initialStateToRestore`. Therefore, I'm adding a gated mechanism that will remove from `initialStatesToRestore` any property handles associated to those cleaned up animation states.

{F1973764319}

For now I only add this integration for IG, since the crash is happening mostly there. I just want to validate that this would indeed mitigate it.

Reviewed By: jettbow

Differential Revision: D67079132

fbshipit-source-id: e3ef250cde5be0c4c1e01448cdbfa4e0b807cce7