Convert SavedReplyGlyphComponentSpec to KComponent
Summary:
## This Diff
This diff converts `SavedReplyGlyphComponentSpec` to KComponent and updates both the MTV and UTV usages (although UTV code will eventually be deleted).

Note: I had to make some changes to core Litho code for this diff. The `StyleCompat` (to be used in Java code) object does not expose a method for setting touch expansion while `Style` (to be used in Kotlin code) does. I added a method to `StyleCompat` to allow us to set this in Java-land, following the example of similar props (e.g. margin).

Differential Revision: D67408743

fbshipit-source-id: eeef53e85cb1a14cbce81d5545d879ad9fd91f94