Add ability to reset initial states on reset()
Summary:
# Context
We have this crash (T209041862) where there seems to be a data inconsistency, because we are trying to restore an initial state, but there is no matching animation state, which leads to a crash.

In previous diffs (D67034475) I have added logging so that we could have a better vision of the lifecycle of these animation states. As we now start to get some data, seems to be a pattern ([evidences](https://docs.google.com/document/d/1RPmW57jBhsuTMVmCID7cRvaB2DRGB7jibPeF00PXFGI/edit?usp=sharing)) that the missing animation state was cleared on [reset](https://fburl.com/code/keimrvj6) of the `TransitionManager`.

This indicates that there is likely a "rogue reset"  that is clearing up animations, but doesn't clear up the initial states to restore. In this diff I add the ability to also reset the `initialStatesToRestore` during `TransitionManager.reset` to understand if the crash is mitigated.

Reviewed By: jettbow

Differential Revision: D67712955

fbshipit-source-id: 64aa70e9564073b6d9446f6809cdd426d2304c52