Supporting size-mode enum in LithoTestRule to enable auto-resize
Summary:
# Problem
As explained in T187041319, using LithoTestRule and you update the size of the root LithoView (e.g. due to a root change or state update), those changes won't be reflected until you call .measure().layout(), since that's when mount() is called in cases we need to resize the LithoView.

# Solution
Added `LithoViewResizeMode` enum and using in `LithoTestRule` to determine auto-resizing.

Reviewed By: adityasharat

Differential Revision: D67935972

fbshipit-source-id: a1fdcf118d6a12d8e87d0a61f4c4e1149b4b51a2