make RenderResult constructor public
Summary:
In Signals, we want to be able to reuse `RenderTree`s across generations if the associated layout result and layout signal hasn't changed. This is particularly of use in determining whether to invalidate a collection's cells as will be seen in upcoming diffs

However, none of `RenderResult`s static creator functions allow this because the decision to reuse a previous `RenderTree` is based off of node/model pointer equality which no longer really works in Signals

Hence, we're opening up the constructor of `RenderResult` to allow callers to instantiate a `RenderResult` with a `RenderTree` of their choice.

Reviewed By: pasqualeanatriello

Differential Revision: D68839596

fbshipit-source-id: 52bf736ad135f20c2647b0efa3a2f0f24f6de1c2