CrossAxisWrapMode.Dynamic documentation - specify dimension is adjusted while scrolling
Summary:
we in IGTX recently tried using CrossAxisWrapMode.Dynamic.  Based on the documentation, we were expecting the height of our H-Scroll Pager to wrap around the largest card, precomputed upfront.  Turns out the expected behavior is not to precompute the heights, but rather to adjust to the tallest as the user scrolls, dynamically.
Wondering if y'all thought the wording in this diff offers a clearer description of its capabilities.  Feel free to disregard if not!

Reviewed By: pentiumao

Differential Revision: D69559580

fbshipit-source-id: 3b35e4e483bbc56a9bee3ec8a7731b4c2ac07f82