Remove ChangesInfo from OnDataRendered
Summary:
## Context

As `ChangesInfo` is designed for Sections, which is not necessary for LazyCollections. In order to move LazyCollections from Sections, the first step is making sure we provide backward capability.
To avoid breaking existing callback of `OnDataRendered`, which requires `ChangesInfo` from the interface [SectionLifecycle::dataRendered](https://fburl.com/code/6eqzwmj7), we have to remove it at least for this moment because our new infra doesn't contain any section info anymore.
Luckily, it turns out no one is using this info, so we can remove it for now and maybe consider adding any necessary change set info later.

Differential Revision: D69786376

fbshipit-source-id: f97e638c11add34a859c7c01176f3a4aed2ec455