Add extra field to identify renders happening during layout
Summary:
Add an extra field to ResolveContext to identify when renders are happening during layout.

We can use this for example, to correctly flag such cases were a component attempts to invoke `Component#measure` during a normal resolve phase. This fixes the initial logic where we were only checking if measure is being called in ResolveContext, since such cases are actually valid if that resolve is executing in the scope of a layout phase

Reviewed By: pentiumao

Differential Revision: D70256384

fbshipit-source-id: af6496b392eb30452e58255d81971e0e98eed8dd