Remove `recyclerTouchEventHandler` from `CollectionRecyclerSpec`
Summary:
We spotted this issue when moving Collections away from Sections, where we should use `Style.onTouch` instead. It's because the prop was introduced(D4788873) for `RecyclerCollectionComponentSpec` as we put the recycler under a wrapper which end up not being able to trigger this event:

https://www.internalfb.com/code/fbsource/[24137574514c53db851094a542471834b428e521][blame]/fbandroid/libraries/components/litho-sections-widget/src/main/java/com/facebook/litho/sections/widget/RecyclerCollectionComponentSpec.java?lines=328-329

But for Collections, we never use the same approach and this property is never used at all.

Reviewed By: astreet

Differential Revision: D70893280

fbshipit-source-id: b9330504672b1519c7caf552ee2a296b795496bf