Back out "Revert D69405559"
Summary:
Reverting the revert.

The playground apps oncall does not own this code or has any expertise to own it. Since the files clearly trace back to litho  it seems clear it belongs to current litho oncall, *if you are not ok with the decision please delete the code* let's not bounce it back and forth and just delete the code.

I'd do it myself but i have no expertise on Android.

Original commit changeset: be759342cb1e

Original Phabricator Diff: D69734440

Differential Revision: D71215235

fbshipit-source-id: 230b88ae155fe8496436180ba634be7b1ba5e981