Ensure Component#measure happens in layout phase
Summary:
This diff aims to restrict the usage of `Component.measure()` API to resolutions that occur during the layout phase. The API should be used in conjunction with `OnCreateLayoutWithSizeSpec` (for Spec) and `RenderWithConstraints` (for KComponents) when component resolution depends on size constraints.

Disallowing `Component.measure()` during the resolve phase is crucial for the following reasons:
- Size constraints is unavailable during the resolve phase
- Layouts created during resolve need to be cached separately, complicating reconciliation and layout caching
- Layout typically occurs after resolve, not vice versa. Hence performing layout during the resolve phase disrupts the natural sequence, leading to unnecessary complexity and potential issues.

By restricting `Component.measure()` to the layout phase, we ensure components remain reusable, maintain a stable progression of phases, and avoid potential performance regressions and hard-to-debug issues.

Reviewed By: pentiumao

Differential Revision: D70625731

fbshipit-source-id: dd81b14ca31c14574d9b9578063d6e37c77349e2