Fix Accessibility Talkback Annoucement Bug
Summary:
### Context
The issue is happening within the ComponentHost
From the history is looks like this was a fix on FB4A for cases in which the text or content description of a button changes after it has been clicked.
  - For example if a button has the text "Active" and after it is clicked it changes to "Inactive" then the new text/content description should be announced.
- Post - https://fb.workplace.com/groups/litho.support/permalink/3779078152413525/

But the problem that this is causing is that each time we insert text or click on a radio button the UI unsets and resets the AX View hierarchy and triggers the announcement of other components in the view.
https://fburl.com/code/y4qtcwxa

## Fix
Before sending the accessibility announcement we should be checking the component that is updating has accessibility focus.

Reviewed By: zielinskimz

Differential Revision: D71347680

fbshipit-source-id: 9b76af56c04e608dbf93929dc20b9cc7cf2eeb47