notify RCAccessibleClickableSpan on keyboard navigation
Summary:
## Stack context

Previously in D66231751 we added an ability to navigate between clickable spans within `RCTextView` using a keyboard. Selected spans are highlighted on navigation with provided `mHighlightColor`.

In this stack we want to add an ability for spans to update themselves when they're selected due to keyboard navigation. This will be used to make span's text color more contrast when it's highlighted.

## This diff
- adds `RCAccessibleClickableSpan#isKeyboardSelected` to be used by inheritors
- updates this property from `RCTextView`

Reviewed By: rooju

Differential Revision: D70779747

fbshipit-source-id: 1a90d9f5bc5dba71f2f9ad0b0d1808381f0e99b1