Set default value of a scroll position to unset
Summary:
There was a crash, from tasks it looks like events, but i wasnt able to reproduce it.
From logs it seems like its crashing when we are trying to set scroll posiiton on null.

 in the spec we are serting the initial value:
https://www.internalfb.com/code/fbsource/[95ee0305a735]/fbandroid/libraries/components/litho-widget/src/main/java/com/facebook/litho/widget/HorizontalScrollComponentSpec.java?lines=70

so i will just use this as the default too and added nullcheck just because mScrollPosition is nullable.

Reviewed By: pentiumao

Differential Revision: D71811606

fbshipit-source-id: a784f92fd5cce736ee2b6121547b700e60dbfdb9