introduce BinderKey for uniquely identifying binders of a render unit
Summary:
We want to be able to specify custom binders without being restricted to the class key identity that RC currently uses to distinguish between binders.

In this diff, we introduce the concept of a `BinderKey` which is a contract that allows for custom binder key types but enforces uniqueness in collections by requiring implementers to implement `equals` and `hashCode`.

The default key type still remains class based

Reviewed By: pasqualeanatriello

Differential Revision: D71897926

fbshipit-source-id: 6ff171f493cdd4afbf779ad5142093e3d06626ea