Introduce TextInput Primitive
Summary:
As per title - this diff adds a primitive implementation of TextInput component.

It tries to keep the new component as close to the old one as possible, reusing whatever possible. Once we run a QE and confirm that it works correctly and doesn't cause any issues, we can refactor it a bit or split single binder into multiple ones to see if it improves perf.

Reviewed By: apowolny

Differential Revision: D71967086

fbshipit-source-id: 180a4cb611b886c537e331c1f8d95be0eaa43f52