fix crash in fullscreen container component
Summary:
Addressing this support post: https://fb.workplace.com/groups/bloks.bugs.regressions/posts/9527788000598116

Litho uses the external layout version to determine if a `setComponentTree(...)` request is skippable. If the layout version is less than the internal version that Litho tracks, the Litho infra throws an exception.
https://www.internalfb.com/code/fbsource/[7409bf79ece7d25ac756dd671a19d0a5e950f8ab]/fbandroid/libraries/components/litho-core/src/main/java/com/facebook/litho/ComponentTree.java?lines=1672-1686

This was causing videos embedded in the fullscreen container component to crash. Mostly because the render tree was measured with the layout version from render state
https://www.internalfb.com/code/fbsource/[c71dd60ccaf7296af7c0bfac4233d2100625d01f]/fbandroid/libraries/bloks-components/bk-component-fullscreencontainer/com/bloks/components/bkcomponentfullscreencontainer/BKBloksComponentFullScreenContainerBinderUtil.kt?lines=84-85
and then measured again when transitioning to fullscreen outside of `RenderState`
https://www.internalfb.com/code/fbsource/[c71dd60ccaf7296af7c0bfac4233d2100625d01f]/fbandroid/libraries/bloks-components/bk-component-fullscreencontainer/com/bloks/components/bkcomponentfullscreencontainer/FullScreenCoordinator.kt?lines=208
https://www.internalfb.com/code/fbsource/[c71dd60ccaf7296af7c0bfac4233d2100625d01f]/fbandroid/libraries/bloks/bloks/src/main/java/com/instagram/common/bloks/Hosts.kt?lines=64%2C67-68%2C70%2C88%2C101

Reviewed By: pasqualeanatriello

Differential Revision: D72331426

fbshipit-source-id: 8c9954b6b53a4412f6cb99215713f3c156b1317b