Make `LegacyMigBottomSheetDialogFragment` Nullsafe (3/n)
Summary:
Made parameter `component` nullable as it is passed to other nullable parameters

## General Context

This diff is part of an effort to make all of our Android apps nullsafe. This will not only help reduce NPEs but allow for smooth conversion from Java to Kotlin. If this is your team’s code, we would appreciate it if you could review our efforts as well as make the rest of your codebase nullsafe.

Reviewed By: pentiumao

Differential Revision: D73800436

fbshipit-source-id: 0cfb698de29298c6a0d1c71b5881edd7785aaf55