Fix TextInput crash during measurement
Summary:
In this diff we're attempting to fix the long-standing issue https://fburl.com/logview/5veuveqs
where we're observing a crash during TextInput measurement. It seems like it happens when setInputTypeAndKeyListenerIfChanged method is called which tries to hide the keyboard and this affects the window size which results in scheduling a layout pass, but before that happens there is a main thread assertion.

The fix is to create a custom Context subclass and override getSystemService method in a way where we return null if someone requests an InputMethodManager. With this we hope to fail this check: https://cs.android.com/android/platform/superproject/main/+/main:frameworks/base/core/java/android/widget/TextView.java;l=7817?q=setInputType
and the keyboard won't be hidden during measure pass.

setInputTypeAndKeyListenerIfChanged is called again during mount with a regular context so the keyboard will be hidden at that point if needed.

Reviewed By: astreet

Differential Revision: D73921811

fbshipit-source-id: dbf0fb4d519772fc2cb39992111e5f23586d515e