use paused LooperMode for all RenderState tests
Summary:
PAUSED LooperMode will become the default soon, and `enqueueStateUpdate_whenTriggered_resolvesAgainWithStateUpdates` was failing with it. So this diff makes the entire class use the PAUSED LooperMode so that the test will continue to pass when the default switches over.

More info in post: https://fb.workplace.com/groups/android.testing.fyi/permalink/2448074072235887/

Reviewed By: cpepin

Differential Revision: D73998808

fbshipit-source-id: d83cfdbe63f9dfab6dce06a296fd5e541270711a